[utils] Split out traversal.py dummy and traversal tests
